tweak to mean pooling. changed .to(float) => .to(token_embeddings.dtype) for correct typ conversion to either torch.float16 or torch.float32

+++ src/model/lightning_model.py
@@ -25,7 +25,7 @@ class OSRSRetrieverModel(pl.LightningModule):
         a single fixed-size sentence/passage vector.
         """
         token_embeddings = model_output[0]  # First element contains hidden states
-        input_mask_expanded = attention_mask.unsqueeze(-1).expand(token_embeddings.size()).float()
+        input_mask_expanded = attention_mask.unsqueeze(-1).expand(token_embeddings.size()).to(token_embeddings.dtype)
         sum_embeddings = torch.sum(token_embeddings * input_mask_expanded, 1)
         sum_mask = torch.clamp(input_mask_expanded.sum(1), min=1e-9)
         return sum_embeddings / sum_mask
